feat(core): Add `getGlobalScope()` method (#9920)
This scope lives in module scope and is applied to _all_ events.

Please review this carefully, as it is important that data is correctly
applied etc. There should be a decent amount of tests covering all of
this, but just to make sure. This was mostly ported/extracted from
node-experimental.
