+++ packages/node-experimental/test/sdk/scope.test.ts
@@ -1,8 +1,7 @@
-import { applyScopeDataToEvent } from '@sentry/core';
-import type { Attachment, Breadcrumb, Client, EventProcessor } from '@sentry/types';
+import { prepareEvent } from '@sentry/core';
+import type { Attachment, Breadcrumb, Client, ClientOptions, EventProcessor } from '@sentry/types';
 import { Scope, getIsolationScope } from '../../src';
-import { getGlobalScope, mergeArray, mergeData, mergePropKeep, mergePropOverwrite } from '../../src/sdk/scope';
-import type { ScopeData } from '../../src/sdk/types';
+import { getGlobalScope } from '../../src/sdk/scope';
 import { mockSdkInit, resetGlobals } from '../helpers/mockSdkInit';
 
 describe('Unit | Scope', () => {
@@ -109,210 +108,41 @@ describe('Unit | Scope', () => {
     expect(scope['_getIsolationScope']()).toBe(customIsolationScope);
   });
 
-  describe('mergeArray', () => {
-    it.each([
-      [[], [], undefined],
-      [undefined, [], undefined],
-      [['a'], [], ['a']],
-      [['a'], ['b', 'c'], ['a', 'b', 'c']],
-      [[], ['b', 'c'], ['b', 'c']],
-      [undefined, ['b', 'c'], ['b', 'c']],
-    ])('works with %s and %s', (a, b, expected) => {
-      const data = { fingerprint: a };
-      mergeArray(data, 'fingerprint', b);
-      expect(data.fingerprint).toEqual(expected);
-    });
-
-    it('does not mutate the original array if no changes are made', () => {
-      const fingerprint = ['a'];
-      const data = { fingerprint };
-      mergeArray(data, 'fingerprint', []);
-      expect(data.fingerprint).toBe(fingerprint);
-    });
-  });
-
-  describe('mergePropKeep', () => {
-    it.each([
-      [{}, {}, {}],
-      [{ a: 'aa' }, {}, { a: 'aa' }],
-      [{ a: 'aa' }, { b: 'bb' }, { a: 'aa', b: 'bb' }],
-      // Does not overwrite existing keys
-      [{ a: 'aa' }, { b: 'bb', a: 'cc' }, { a: 'aa', b: 'bb' }],
-    ])('works with %s and %s', (a, b, expected) => {
-      const data = { tags: a } as unknown as ScopeData;
-      mergePropKeep(data, 'tags', b);
-      expect(data.tags).toEqual(expected);
-    });
-
-    it('does not deep merge', () => {
-      const data = {
-        contexts: {
-          app: { app_version: 'v1' },
-          culture: { display_name: 'name1' },
-        },
-      } as unknown as ScopeData;
-      mergePropKeep(data, 'contexts', {
-        os: { name: 'os1' },
-        app: { app_name: 'name1' },
-      });
-      expect(data.contexts).toEqual({
-        os: { name: 'os1' },
-        culture: { display_name: 'name1' },
-        app: { app_version: 'v1' },
-      });
-    });
-
-    it('does not mutate the original object if no changes are made', () => {
-      const tags = { a: 'aa' };
-      const data = { tags } as unknown as ScopeData;
-      mergePropKeep(data, 'tags', {});
-      expect(data.tags).toBe(tags);
-    });
-  });
-
-  describe('mergePropOverwrite', () => {
-    it.each([
-      [{}, {}, {}],
-      [{ a: 'aa' }, {}, { a: 'aa' }],
-      [{ a: 'aa' }, { b: 'bb' }, { a: 'aa', b: 'bb' }],
-      // overwrites existing keys
-      [{ a: 'aa' }, { b: 'bb', a: 'cc' }, { a: 'cc', b: 'bb' }],
-    ])('works with %s and %s', (a, b, expected) => {
-      const data = { tags: a } as unknown as ScopeData;
-      mergePropOverwrite(data, 'tags', b);
-      expect(data.tags).toEqual(expected);
-    });
-
-    it('does not deep merge', () => {
-      const data = {
-        contexts: {
-          app: { app_version: 'v1' },
-          culture: { display_name: 'name1' },
-        },
-      } as unknown as ScopeData;
-      mergePropOverwrite(data, 'contexts', {
-        os: { name: 'os1' },
-        app: { app_name: 'name1' },
-      });
-      expect(data.contexts).toEqual({
-        os: { name: 'os1' },
-        culture: { display_name: 'name1' },
-        app: { app_name: 'name1' },
-      });
-    });
-
-    it('does not mutate the original object if no changes are made', () => {
-      const tags = { a: 'aa' };
-      const data = { tags } as unknown as ScopeData;
-      mergePropOverwrite(data, 'tags', {});
-      expect(data.tags).toBe(tags);
-    });
-  });
-
-  describe('mergeData', () => {
-    it('works with empty data', () => {
-      const data1: ScopeData = {
-        eventProcessors: [],
-        breadcrumbs: [],
-        user: {},
-        tags: {},
-        extra: {},
-        contexts: {},
-        attachments: [],
-        propagationContext: { spanId: '1', traceId: '1' },
-        sdkProcessingMetadata: {},
-        fingerprint: [],
-      };
-      const data2: ScopeData = {
-        eventProcessors: [],
-        breadcrumbs: [],
-        user: {},
-        tags: {},
-        extra: {},
-        contexts: {},
-        attachments: [],
-        propagationContext: { spanId: '1', traceId: '1' },
-        sdkProcessingMetadata: {},
-        fingerprint: [],
-      };
-      mergeData(data1, data2);
-      expect(data1).toEqual({
-        eventProcessors: [],
-        breadcrumbs: [],
-        user: {},
-        tags: {},
-        extra: {},
-        contexts: {},
-        attachments: [],
-        propagationContext: { spanId: '1', traceId: '1' },
-        sdkProcessingMetadata: {},
-        fingerprint: [],
-      });
-    });
-
-    it('merges data correctly', () => {
-      const attachment1 = { filename: '1' } as Attachment;
-      const attachment2 = { filename: '2' } as Attachment;
-      const attachment3 = { filename: '3' } as Attachment;
-
-      const breadcrumb1 = { message: '1' } as Breadcrumb;
-      const breadcrumb2 = { message: '2' } as Breadcrumb;
-      const breadcrumb3 = { message: '3' } as Breadcrumb;
-
-      const eventProcessor1 = ((a: unknown) => null) as EventProcessor;
-      const eventProcessor2 = ((b: unknown) => null) as EventProcessor;
-      const eventProcessor3 = ((c: unknown) => null) as EventProcessor;
-
-      const data1: ScopeData = {
-        eventProcessors: [eventProcessor1],
-        breadcrumbs: [breadcrumb1],
-        user: { id: '1', email: 'test@example.com' },
-        tags: { tag1: 'aa', tag2: 'aa' },
-        extra: { extra1: 'aa', extra2: 'aa' },
-        contexts: { os: { name: 'os1' }, culture: { display_name: 'name1' } },
-        attachments: [attachment1],
-        propagationContext: { spanId: '1', traceId: '1' },
-        sdkProcessingMetadata: { aa: 'aa', bb: 'aa' },
-        fingerprint: ['aa', 'bb'],
-      };
-      const data2: ScopeData = {
-        eventProcessors: [eventProcessor2, eventProcessor3],
-        breadcrumbs: [breadcrumb2, breadcrumb3],
-        user: { id: '2', name: 'foo' },
-        tags: { tag2: 'bb', tag3: 'bb' },
-        extra: { extra2: 'bb', extra3: 'bb' },
-        contexts: { os: { name: 'os2' } },
-        attachments: [attachment2, attachment3],
-        propagationContext: { spanId: '2', traceId: '2' },
-        sdkProcessingMetadata: { bb: 'bb', cc: 'bb' },
-        fingerprint: ['cc'],
-      };
-      mergeData(data1, data2);
-      expect(data1).toEqual({
-        eventProcessors: [eventProcessor1, eventProcessor2, eventProcessor3],
-        breadcrumbs: [breadcrumb1, breadcrumb2, breadcrumb3],
-        user: { id: '2', name: 'foo', email: 'test@example.com' },
-        tags: { tag1: 'aa', tag2: 'bb', tag3: 'bb' },
-        extra: { extra1: 'aa', extra2: 'bb', extra3: 'bb' },
-        contexts: { os: { name: 'os2' }, culture: { display_name: 'name1' } },
-        attachments: [attachment1, attachment2, attachment3],
-        propagationContext: { spanId: '2', traceId: '2' },
-        sdkProcessingMetadata: { aa: 'aa', bb: 'bb', cc: 'bb' },
-        fingerprint: ['aa', 'bb', 'cc'],
-      });
-    });
-  });
-
-  describe('applyToEvent', () => {
-    it('works without any data', async () => {
+  describe('prepareEvent', () => {
+    it('works without any scope data', async () => {
       mockSdkInit();
+
+      const eventProcessor = jest.fn((a: unknown) => a) as EventProcessor;
 
       const scope = new Scope();
 
       const event = { message: 'foo' };
-      applyScopeDataToEvent(event, scope.getScopeData());
 
-      expect(event).toEqual({
+      const options = {} as ClientOptions;
+      const client = {
+        getEventProcessors() {
+          return [eventProcessor];
+        },
+      } as Client;
+      const processedEvent = await prepareEvent(
+        options,
+        event,
+        {
+          integrations: [],
+        },
+        scope,
+        client,
+      );
+
+      expect(eventProcessor).toHaveBeenCalledWith(processedEvent, {
+        integrations: [],
+        // no attachments are added to hint
+      });
+
+      expect(processedEvent).toEqual({
+        timestamp: expect.any(Number),
+        event_id: expect.any(String),
+        environment: 'production',
         message: 'foo',
         sdkProcessingMetadata: {
           propagationContext: {
@@ -335,6 +165,10 @@ describe('Unit | Scope', () => {
       const eventProcessor2 = jest.fn((b: unknown) => b) as EventProcessor;
       const eventProcessor3 = jest.fn((c: unknown) => c) as EventProcessor;
 
+      const attachment1 = { filename: '1' } as Attachment;
+      const attachment2 = { filename: '2' } as Attachment;
+      const attachment3 = { filename: '3' } as Attachment;
+
       const scope = new Scope();
       scope.update({
         user: { id: '1', email: 'test@example.com' },
@@ -346,6 +180,7 @@ describe('Unit | Scope', () => {
       });
       scope.addBreadcrumb(breadcrumb1);
       scope.addEventProcessor(eventProcessor1);
+      scope.addAttachment(attachment1);
 
       const globalScope = getGlobalScope();
       const isolationScope = getIsolationScope();
@@ -353,16 +188,39 @@ describe('Unit | Scope', () => {
       globalScope.addBreadcrumb(breadcrumb2);
       globalScope.addEventProcessor(eventProcessor2);
       globalScope.setSDKProcessingMetadata({ aa: 'aa' });
+      globalScope.addAttachment(attachment2);
 
       isolationScope.addBreadcrumb(breadcrumb3);
       isolationScope.addEventProcessor(eventProcessor3);
-      globalScope.setSDKProcessingMetadata({ bb: 'bb' });
+      isolationScope.setSDKProcessingMetadata({ bb: 'bb' });
+      isolationScope.addAttachment(attachment3);
 
       const event = { message: 'foo', breadcrumbs: [breadcrumb4], fingerprint: ['dd'] };
 
-      applyScopeDataToEvent(event, scope.getScopeData());
+      const options = {} as ClientOptions;
+      const processedEvent = await prepareEvent(
+        options,
+        event,
+        {
+          integrations: [],
+        },
+        scope,
+      );
 
-      expect(event).toEqual({
+      expect(eventProcessor1).toHaveBeenCalledTimes(1);
+      expect(eventProcessor2).toHaveBeenCalledTimes(1);
+      expect(eventProcessor3).toHaveBeenCalledTimes(1);
+
+      // Test that attachments are correctly merged
+      expect(eventProcessor1).toHaveBeenCalledWith(processedEvent, {
+        integrations: [],
+        attachments: [attachment2, attachment3, attachment1],
+      });
+
+      expect(processedEvent).toEqual({
+        timestamp: expect.any(Number),
+        event_id: expect.any(String),
+        environment: 'production',
         message: 'foo',
         user: { id: '1', email: 'test@example.com' },
         tags: { tag1: 'aa', tag2: 'aa' },
@@ -379,37 +237,6 @@ describe('Unit | Scope', () => {
           },
         },
       });
-    });
-  });
-
-  describe('getAttachments', () => {
-    it('works without any data', async () => {
-      mockSdkInit();
-
-      const scope = new Scope();
-
-      const actual = scope.getAttachments();
-      expect(actual).toEqual([]);
-    });
-
-    it('merges attachments data', async () => {
-      mockSdkInit();
-
-      const attachment1 = { filename: '1' } as Attachment;
-      const attachment2 = { filename: '2' } as Attachment;
-      const attachment3 = { filename: '3' } as Attachment;
-
-      const scope = new Scope();
-      scope.addAttachment(attachment1);
-
-      const globalScope = getGlobalScope();
-      const isolationScope = getIsolationScope();
-
-      globalScope.addAttachment(attachment2);
-      isolationScope.addAttachment(attachment3);
-
-      const actual = scope.getAttachments();
-      expect(actual).toEqual([attachment2, attachment3, attachment1]);
     });
   });
 });
